# apps/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/billing/plans": {
      "get": {
        "operationId": "BillingController_getPlans",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PlanPriceDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List plan prices from Stripe (protected)",
        "tags": [
          "billing"
        ]
      }
    },
    "/billing/checkout-session": {
      "post": {
        "operationId": "BillingController_createCheckoutSession",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCheckoutSessionDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BillingRedirectResponseDto"
                }
              }
            }
          },
          "409": {
            "description": "SUBSCRIPTION_ALREADY_ACTIVE: the tenant already has a live subscription."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a Stripe Checkout session (protected)",
        "tags": [
          "billing"
        ]
      }
    },
    "/billing/portal-session": {
      "post": {
        "operationId": "BillingController_createPortalSession",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BillingRedirectResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a Stripe billing portal session (protected)",
        "tags": [
          "billing"
        ]
      }
    },
    "/billing/plan-change": {
      "get": {
        "operationId": "BillingController_getPlanChange",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlanChangeStatusResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get plan change options (protected)",
        "tags": [
          "billing"
        ]
      },
      "post": {
        "operationId": "BillingController_requestPlanChange",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RequestPlanChangeDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlanChangeResultDto"
                }
              }
            }
          },
          "409": {
            "description": "PLAN_CHANGE_NOT_ALLOWED, or PLAN_CHANGE_IMPACT_NOT_ACKNOWLEDGED with `details.items` when a downgrade loses something and `acknowledgeLosses` is not true."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Request a plan change (protected)",
        "tags": [
          "billing"
        ]
      }
    },
    "/billing/plan-change/pending": {
      "delete": {
        "operationId": "BillingController_releasePendingPlanChange",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReleasePlanChangeResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Cancel the scheduled plan change (protected)",
        "tags": [
          "billing"
        ]
      }
    },
    "/billing/subscription": {
      "get": {
        "operationId": "BillingController_getSubscription",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/SubscriptionResponseDto"
                    },
                    {
                      "type": "null"
                    }
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get the tenant current subscription (protected)",
        "tags": [
          "billing"
        ]
      }
    },
    "/billing/subscription/price": {
      "get": {
        "operationId": "BillingController_getSubscriptionPrice",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubscriptionPriceResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get the price the tenant subscription is billed (protected)",
        "tags": [
          "billing"
        ]
      }
    },
    "/billing/invoices": {
      "get": {
        "operationId": "BillingController_getInvoices",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvoiceListResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get the tenant paid invoice history (protected)",
        "tags": [
          "billing"
        ]
      }
    },
    "/tenants/{tenantId}/plan-change-impact": {
      "get": {
        "operationId": "PlanChangeImpactController_getImpact",
        "parameters": [
          {
            "name": "tenantId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "target",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "basico",
                "pro"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlanChangeImpactResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid bearer token."
          },
          "403": {
            "description": "Tenant does not belong to the caller."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Preview what a plan change removes (protected)",
        "tags": [
          "billing"
        ]
      }
    },
...
  },
  "info": {
    "title": "Lattiz API",
    "description": "Lattiz resource server. Auth is handled by Supabase; this API only validates Supabase-issued JWTs (JWKS / ES256, audience `authenticated`).",
    "version": "0.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "Supabase access token (JWT)."
      }
    },
    "schemas": {
...
      "PlanPriceDto": {
        "type": "object",
        "properties": {
          "plan": {
            "enum": [
              "basico",
              "pro"
            ],
            "type": "string"
          },
          "period": {
            "enum": [
              "monthly",
              "annual"
            ],
            "type": "string"
          },
          "amount": {
            "type": "number",
            "description": "Price per billing period, in cents (minor currency unit)."
          },
          "currency": {
            "type": "string",
            "description": "ISO 4217, lowercase (e.g. `mxn`)."
          },
          "lookupKey": {
            "type": "string"
          }
        },
        "required": [
          "plan",
          "period",
          "amount",
          "currency",
          "lookupKey"
        ]
      },
      "CreateCheckoutSessionDto": {
        "type": "object",
        "properties": {
          "plan": {
            "enum": [
              "basico",
              "pro"
            ],
            "type": "string"
          },
          "period": {
            "enum": [
              "monthly",
              "annual"
            ],
            "type": "string"
          }
        },
        "required": [
          "plan",
          "period"
        ]
      },
      "BillingRedirectResponseDto": {
        "type": "object",
        "properties": {
          "url": {
            "type": "string"
          }
        },
        "required": [
          "url"
        ]
      },
      "PendingPlanChangeDto": {
        "type": "object",
        "properties": {
          "targetPlan": {
            "enum": [
              "basico",
              "pro"
            ],
            "type": "string"
          },
          "effectiveAt": {
            "type": "string",
            "description": "ISO date the scheduled change takes effect."
          }
        },
        "required": [
          "targetPlan",
          "effectiveAt"
        ]
      },
      "PlanChangeOptionDto": {
        "type": "object",
        "properties": {
          "targetPlan": {
            "enum": [
              "basico",
              "pro"
            ],
            "type": "string"
          },
          "direction": {
            "description": "`none` when the target is the current plan or the current plan is unknown.",
            "enum": [
              "upgrade",
              "downgrade",
              "none"
            ],
            "type": "string"
          },
          "effective": {
            "nullable": true,
            "description": "Upgrades apply now (prorated); downgrades at the end of the current billing period.",
            "enum": [
              "immediate",
              "period_end",
              null
            ],
            "type": "string"
          },
          "effectiveAt": {
            "type": "string",
            "nullable": true,
            "description": "ISO date a downgrade would take effect; null for upgrades."
          },
          "allowed": {
            "type": "boolean"
          },
          "blockers": {
            "type": "array",
            "items": {
              "type": "string",
              "enum": [
                "NO_ACTIVE_SUBSCRIPTION",
                "SUBSCRIPTION_NOT_ACTIVE",
                "SUBSCRIPTION_CANCELING",
                "SAME_PLAN",
                "PLAN_CHANGE_PENDING",
                "UNSUPPORTED_SUBSCRIPTION",
                "DOWNGRADE_DOMAIN_ABOVE_BASIC_CAP",
                "DOWNGRADE_PRICE_UNAVAILABLE"
              ]
            }
          }
        },
        "required": [
          "targetPlan",
          "direction",
          "effective",
          "effectiveAt",
          "allowed",
          "blockers"
        ]
      },
      "PlanChangeStatusResponseDto": {
        "type": "object",
        "properties": {
          "currentPlan": {
            "nullable": true,
            "description": "Derived from the subscription's price; null without a subscription or for an unknown price.",
            "enum": [
              "basico",
              "pro",
              null
            ],
            "type": "string"
          },
          "billingPeriod": {
            "nullable": true,
            "enum": [
              "monthly",
              "annual",
              null
            ],
            "type": "string"
          },
          "pending": {
            "nullable": true,
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/PendingPlanChangeDto"
              }
            ]
          },
          "options": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PlanChangeOptionDto"
            }
          }
        },
        "required": [
          "currentPlan",
          "billingPeriod",
          "pending",
          "options"
        ]
      },
      "RequestPlanChangeDto": {
        "type": "object",
        "properties": {
          "targetPlan": {
            "enum": [
              "basico",
              "pro"
            ],
            "type": "string"
          },
          "acknowledgeLosses": {
            "type": "boolean",
            "description": "Required (true) when GET /tenants/:tenantId/plan-change-impact lists any loss;\notherwise the API answers 409 PLAN_CHANGE_IMPACT_NOT_ACKNOWLEDGED with the impact."
          }
        },
        "required": [
          "targetPlan"
        ]
      },
      "PlanChangeResultDto": {
        "type": "object",
        "properties": {
          "kind": {
            "enum": [
              "redirect",
              "scheduled"
            ],
            "type": "string"
          },
          "url": {
            "type": "string",
            "nullable": true
          },
          "effectiveAt": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "kind",
          "url",
          "effectiveAt"
        ]
      },
      "ReleasePlanChangeResponseDto": {
        "type": "object",
        "properties": {
          "released": {
            "type": "boolean"
          }
        },
        "required": [
          "released"
        ]
      },
      "SubscriptionResponseDto": {
        "type": "object",
        "properties": {
          "plan": {
            "enum": [
              "basico",
              "pro"
            ],
            "type": "string"
          },
          "billingPeriod": {
            "enum": [
              "monthly",
              "annual"
            ],
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "active",
              "trialing",
              "past_due",
              "canceled",
              "incomplete",
              "incomplete_expired",
              "unpaid",
              "paused"
            ]
          },
          "currentPeriodEnd": {
            "type": "string",
            "nullable": true
          },
          "cancelAt": {
            "type": "string",
            "nullable": true,
            "description": "ISO date when the subscription ends, if a cancellation is scheduled; null otherwise."
          },
          "cancelAtPeriodEnd": {
            "type": "boolean"
          }
        },
        "required": [
          "plan",
          "billingPeriod",
          "status",
          "currentPeriodEnd",
          "cancelAt",
          "cancelAtPeriodEnd"
        ]
      },
      "SubscriptionPriceDto": {
        "type": "object",
        "properties": {
          "amount": {
            "type": "number",
            "description": "Price per billing period, in cents (minor currency unit)."
          },
          "currency": {
            "type": "string",
            "description": "ISO 4217, lowercase (e.g. `mxn`)."
          },
          "period": {
            "enum": [
              "monthly",
              "annual"
            ],
            "type": "string"
          }
        },
        "required": [
          "amount",
          "currency",
          "period"
        ]
      },
      "SubscriptionPriceResponseDto": {
        "type": "object",
        "properties": {
          "price": {
            "nullable": true,
            "description": "Null when the tenant has never subscribed.",
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/SubscriptionPriceDto"
              }
            ]
          }
        },
        "required": [
          "price"
        ]
      },
      "InvoiceDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "date": {
            "type": "number",
            "description": "Unix timestamp (seconds)."
          },
          "amountPaid": {
            "type": "number",
            "description": "Amount paid, in cents."
          },
          "currency": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "periodStart": {
            "type": "number"
          },
          "periodEnd": {
            "type": "number"
          },
          "invoicePdf": {
            "type": "string",
            "nullable": true
          },
          "hostedInvoiceUrl": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "id",
          "date",
          "amountPaid",
          "currency",
          "status",
          "periodStart",
          "periodEnd",
          "invoicePdf",
          "hostedInvoiceUrl",
          "description"
        ]
      },
      "InvoiceListResponseDto": {
        "type": "object",
        "properties": {
          "invoices": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/InvoiceDto"
            }
          }
        },
        "required": [
          "invoices"
        ]
      },
      "PlanChangeImpactItemDto": {
        "type": "object",
        "properties": {
          "kind": {
            "enum": [
              "template_loss",
              "analytics",
              "domain_renewal_cap"
            ],
            "type": "string"
          },
          "templateId": {
            "type": "string",
            "nullable": true
          },
          "templateName": {
            "type": "string",
            "nullable": true
          },
          "effectiveAt": {
            "type": "string",
            "nullable": true,
            "description": "ISO date the loss takes effect (period end for a downgrade)."
          }
        },
        "required": [
          "kind",
          "templateId",
          "templateName",
          "effectiveAt"
        ]
      },
      "PlanChangeImpactResponseDto": {
        "type": "object",
        "properties": {
          "target": {
            "enum": [
              "basico",
              "pro"
            ],
            "type": "string"
          },
          "items": {
            "description": "Empty when nothing is lost. A non-empty list must be acknowledged to schedule the change.",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PlanChangeImpactItemDto"
            }
          }
        },
        "required": [
          "target",
          "items"
        ]
      },
...
    }
  }
}